requirements.txt
Flask==3.0.0
supabase
python-dotenv
Flask-Session
